requirements.txt
annotated-types==0.6.0
attrs==23.2.0
Babel==2.14.0
fluent.runtime==0.4.0
fluent.syntax==0.19.0
pydantic==2.6.1
pydantic_core==2.16.2
pytz==2024.1
typing_extensions==4.9.0
